Add: Support for batching/queuing events
Add: Support for batching/queuing events
Add: Helper functions
Add: Requirements.txt for examples
Refactor: Examples README.md
Refactor: License
Refactor: Add apscheduler as install requires the package
Bump version to 1.3.0
