+++ LICENSE
@@ -1,4 +1,4 @@
-﻿Copyright (c) 2019 Moesif, Inc
+﻿Copyright (c) 2020 Moesif, Inc
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
