+++ examples/falcon/requirements.txt
@@ -1,0 +1,4 @@
+moesifwsgi
+falcon
+gunicorn
+falcon-multipart
