+++ examples/flask/requirements.txt
@@ -1,0 +1,2 @@
+moesifwsgi
+Flask
