+++ setup.py
@@ -28,7 +28,7 @@ setup(
     # Versions should comply with PEP440.  For a discussion on single-sourcing
     # the version across setup.py and the project code, see
     # https://packaging.python.org/en/latest/single_source_version.html
-    version='1.2.6',
+    version='1.3.0',
 
     description='Moesif Middleware for Python WSGI based flatforms (Flask, Bottle & Others)',
     long_description=long_description,
@@ -82,7 +82,7 @@ setup(
     # your project is installed. For an analysis of "install_requires" vs pip's
     # requirements files see:
     # https://packaging.python.org/en/latest/requirements.html
-    install_requires=['requests', 'isodatetimehandler', 'moesifapi', 'moesifpythonrequest'],
+    install_requires=['requests', 'isodatetimehandler', 'moesifapi', 'moesifpythonrequest', 'apscheduler'],
 
     # List additional groups of dependencies here (e.g. development
     # dependencies). You can install these using the following syntax,
